+++ packages/react-ui/src/theme/sections/components/forms.js
@@ -248,7 +248,7 @@ export const formsOverrides = {
           },
           '&.Mui-readOnly': {
             '&.Mui-focused': {
-              backgroundColor: theme.palette.background.default
+              backgroundColor: theme.palette.default.background
             }
           }
         },
@@ -266,10 +266,10 @@ export const formsOverrides = {
             backgroundColor: theme.palette.default.background
           },
           '&.Mui-readOnly': {
-            backgroundColor: theme.palette.background.default,
+            backgroundColor: theme.palette.default.background,
 
             '&.Mui-focused': {
-              backgroundColor: theme.palette.background.default
+              backgroundColor: theme.palette.default.background
             }
           },
           '& .MuiOutlinedInput-notchedOutline': {